pyproject.toml
[project]
name = "google-docs-mcp"
version = "1.0.0"
description = "Google Docs MCP Server for reading, creating, and editing Google Documents"
dependencies = [
    "fastmcp",
    "google-auth",
    "google-auth-oauthlib",
    "google-api-python-client",
    "python-dotenv",
    "mcp-mapped-resource-lib",
    "docker",
]
requires-python = ">=3.10"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
google-docs-mcp = "google_docs_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/google_docs_mcp"]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
]
